requirements.txt
# Ядро для роботи з браузером
playwright==1.40.0

# Валідація даних та моделі
pydantic==2.4.1

# Професійне логування (кольорові повідомлення в консолі)
loguru==0.7.2

# Додаткова бібліотека для роботи з Excel (якщо захочеш зберігати в .xlsx)
#pandas==2.1.3
openpyxl==3.1.2

fake-useragent
pyyaml
requests

#pip install -r requirements.txt
#playwright install chromium